api/requirements.txt
fastapi==0.116.2
uvicorn==0.25.0
python-dotenv==1.1.1
psycopg2-binary==2.9.11
sqlalchemy==2.0.43
pydantic==2.7.2
supabase==2.22.0
